pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "d-tools-cloud-mcp-server"
version = "0.1.0"
description = "Model Context Protocol server for D-Tools Cloud API"
readme = "README.md"
requires-python = ">=3.13"
authors = [{name = "James Young"}]
dependencies = [
    "anthropic>=0.96.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
d-tools-mcp = "dtools_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dtools_mcp"]
